requirements.txt
Django
djangorestframework
requests
beautifulsoup4
pytest
pytest-django
